Add `WorldManager`, `GameContext` and change game structure a bit
This is a rather large commit. It's the first working version after
restructuring the whole application.

First, the `WorldManager` will manage the world and world view. It
will dynamically load chunks and update the view. It implements clone
and behaves like an `Rc`; everyone can own a `WorldManager` to get
access to the world.

Second, the `GameContext` owns the facade and stores the game config.
It's merely a utility helper that can be stored in every component of
the game to get quick access to the facade and the config.

I'm really not sure about all this. And it needs some cleanup still.